requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0
peft>=0.7.0
accelerate>=0.24.0
safetensors>=0.4.0

# Experiment Tracking
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Evaluation Metrics
scikit-learn>=1.3.0
evaluate>=0.4.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0

# Notebook Support
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Optional: Memory Optimization
bitsandbytes>=0.41.0